+++ components/private/Snapshot/BUILD
@@ -14,6 +14,7 @@
 
 load(
     "//:material_components_ios.bzl",
+    "mdc_extension_objc_library",
     "mdc_objc_library",
     "mdc_public_objc_library",
 )
@@ -38,19 +39,15 @@ swift_library(
     ],
 )
 
-mdc_objc_library(
-    name = "Supplemental",
+mdc_extension_objc_library(
+    name = "SnapshotUtilities",
     testonly = 1,
-    srcs = glob(["src/Supplemental/*.m"]),
-    hdrs = glob(["src/Supplemental/*.h"]),
     visibility = ["//visibility:public"],
 )
 
-mdc_objc_library(
-    name = "Implementation",
+mdc_extension_objc_library(
+    name = "SnapshotTestCase",
     testonly = 1,
-    srcs = glob(["src/Implementation/*.m"]),
-    hdrs = glob(["src/Implementation/*.h"]),
     visibility = ["//visibility:private"],
     deps = [
         "@ios_snapshot_test_case//:SnapshotTestCase",
@@ -61,9 +58,9 @@ mdc_public_objc_library(
     name = "Snapshot",
     testonly = 1,
     deps = [
-        ":Implementation",
+        ":SnapshotTestCase",
         ":SnapshotSourceDummies",
         ":SnapshotSwiftLib",
-        ":Supplemental",
+        ":SnapshotUtilities",
     ],
 )
