+++ components/private/Snapshot/src/MaterialSnapshot.h
@@ -12,6 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#import "Implementation/MDCSnapshotTestCase.h"
-#import "Supplemental/UIImage+MDCSnapshot.h"
-#import "Supplemental/UIView+MDCSnapshot.h"
+#import "SnapshotTestCase/MDCSnapshotTestCase.h"
+#import "SnapshotUtilities/UIImage+MDCSnapshot.h"
+#import "SnapshotUtilities/UIView+MDCSnapshot.h"
