+++ components/private/Snapshot/src/SnapshotUtilities/UIImage+MDCSnapshot.h
