+++ components/private/Snapshot/src/SnapshotUtilities/UIImage+MDCSnapshot.m
