+++ components/private/Snapshot/src/SnapshotUtilities/UIView+MDCSnapshot.h
